+++ lib/marskal_search/help.rb
@@ -1,13 +1,12 @@
 class MarskalSearch
 
-#NOTE LEft off handling "\n" characters
-
-  def help(p_key = nil)
-    MarskalSearch::Help.commands(p_key)
+  def help(p_key, p_wrap)
+    MarskalSearch::Help.commands(p_key, p_wrap)
   end
 
   class Help
     @@help_content = nil
+    @wrap = true
 
 
     def self.load_content
@@ -15,18 +14,20 @@ class MarskalSearch
       @@help_content = YAML.load(l_erb_file.result(binding)).deep_symbolize_keys()
     end
 
-    def self.commands(p_key = nil)
+    def self.commands(p_key, p_wrap= false)
+      @wrap = p_wrap
       return "General Instructions Coming Soon" if p_key.nil?
-      load_content() if @@help_content.nil? || true
+      load_content() #TODO: put this check back when out of alpha mode if @@help_content.nil?
 
       p_key = p_key.to_sym
 
-      l_output = build_help_string(p_key)
+      l_output = command_help_template(p_key)
       puts l_output
 
       return
     end
 
+    #show all valid keys in synopsis
     def self.default_synopsis(p_key)
       l_text = ''
       VALID_KEYS[p_key][:valid].each do |l_key|
@@ -35,15 +36,18 @@ class MarskalSearch
       l_text
     end
 
-
     private
 
+    #format text for an 80 character wide screen
     def self.format_for_terminal(p_block_text, p_tabs = "\t\t")
+      l_max_length = @wrap ? 79 : 99999999
+      #note we use our own notation "+\n" indicates we are concatenating two lines, so we should replace the \n with a space
       l_text = p_block_text.to_s.split(/\s!\n!\t/).join(' ')
       l_text.gsub!("\\n", "\n")
+      l_text.gsub!("+\n", ' ')
 
       p_tabs = p_tabs.gsub("\t", '  ')
-      l_max_length = 79 -p_tabs.length
+      l_max_length = l_max_length - p_tabs.length
 
       l_formatted_str = ''
       until l_text.blank? do
@@ -61,16 +65,28 @@ class MarskalSearch
           l_last_space = l_line.length if l_last_space.nil? || l_text.length <= l_max_length #chedk if we are on last line
         end
 
-
-
         l_new_str = l_text.slice!(0, l_last_space+1)
         l_formatted_str += "#{p_tabs}#{l_new_str}#{l_char_to_add}"
       end
-      zz = l_formatted_str.gsub("\\t", '  ')
-      zz
+      l_formatted_str.gsub("\\t", '  ')
     end
 
-    def self.build_help_string(p_key_symbol)
+    #format array of options for 80 character terminal mode
+    def self.format_options_for_terminal(p_options)
+      l_result = ''
+      l_ctr = 0
+      p_options.each do |k, p_option|
+        l_result += "\n\n" unless l_ctr == 0
+        l_result  += format_for_terminal(p_option[:name])
+        l_result  += "#{format_for_terminal(p_option[:description], "\t\t\t")}"
+        l_ctr += 1
+      end
+      l_result
+    end
+
+
+    #template for help for a command
+    def self.command_help_template(p_key_symbol)
       p_content = @@help_content[p_key_symbol]
       p_key = p_content[:name]
       <<-OUTSTR
@@ -84,41 +100,28 @@ SYNOPSIS:
     #{p_key} - #{p_content[:synopsis]}
 
     Getting Value:
-      <marskal_object>.#{p_content[:getter]}
+      #{p_content[:getter]}
 
     Setting Value:
-      <marskal_object>.set_#{p_content[:setter]}
+      set_#{p_content[:setter]}
 
 DESCRIPTION:
 
 #{format_for_terminal(p_content[:description])}
 
+OPTIONS
+
+#{format_options_for_terminal(p_content[:options])}
+
 EXAMPLES:
 
-#{format_for_terminal(p_content[:example])}
-
-
+#{format_for_terminal(p_content[:examples])}
 --------------------- END Help #{p_key} ---------------------
 
       OUTSTR
 
     end
-#
-# HELP = { }
-#
-# HELP[:wrap_column] = {}
-# HELP[:wrap_column][:synopsis] = <<-EOS
-#     EOS
-#
-# HELP[:wrap_column][:description] = <<-EOS
-#     It is a long established fact that a reader will be distracted by the readable content of a page when looking at its
-#     layout. The point of using Lorem Ipsum is that it has a more-or-less normal distribution of letters, as opposed to
-#     using 'Content here, content here', making it look like readable English. Many desktop publishing packages and web
-#     page editors now use Lorem Ipsum as their default model text, and a search for 'lorem ipsum' will uncover many web
-#     sites still in their infancy. Various versions have evolved over the years, sometimes by accident, sometimes on
-#     purpose (injected humour and the like).
-#     EOS
-#
+
   end #help
 
 end #marskalsearch